Remove redundant div and share data between tables
