+++ sections/futures/UserInfo/OpenOrdersTable.tsx
@@ -9,6 +9,7 @@ import styled from 'styled-components';
 
 import Badge from 'components/Badge';
 import Currency from 'components/Currency';
+import { DesktopOnlyView, MobileOrTabletView } from 'components/Media';
 import Table from 'components/Table';
 import PositionType from 'components/Text/PositionType';
 import Connector from 'containers/Connector';
@@ -16,9 +17,11 @@ import TransactionNotifier from 'containers/TransactionNotifier';
 import useGetNextPriceDetails from 'queries/futures/useGetNextPriceDetails';
 import { currentMarketState, openOrdersState } from 'store/futures';
 import { gasSpeedState, walletAddressState } from 'store/wallet';
+import { GridDivCenteredRow } from 'styles/common';
 import { formatCurrency } from 'utils/formatters/number';
 import { getDisplayAsset, MarketKeyByAsset, FuturesMarketAsset } from 'utils/futures';
 
+import OrderDrawer from '../MobileTrade/drawers/OrderDrawer';
 import { PositionSide } from '../types';
 
 const OpenOrdersTable: React.FC = () => {
@@ -35,6 +38,7 @@ const OpenOrdersTable: React.FC = () => {
 	const { handleRefetch } = useRefetchContext();
 
 	const [action, setAction] = React.useState<'' | 'cancel' | 'execute'>('');
+	const [selectedOrder, setSelectedOrder] = React.useState<any>();
 
 	const ethGasPriceQuery = useEthGasPriceQuery();
 
@@ -95,98 +99,180 @@ const OpenOrdersTable: React.FC = () => {
 	}, [openOrders, nextPriceDetails?.currentRoundId, synthsMap]);
 
 	return (
-		<StyledTable
-			data={data}
-			highlightRowsOnHover
-			showPagination
-			columns={[
-				{
-					Header: (
-						<StyledTableHeader>
-							{t('futures.market.user.open-orders.table.market-type')}
-						</StyledTableHeader>
-					),
-					accessor: 'market',
-					Cell: (cellProps: CellProps<any>) => {
-						return (
-							<MarketContainer>
-								<IconContainer>
-									<StyledCurrencyIcon currencyKey={cellProps.row.original.marketKey} />
-								</IconContainer>
-								<StyledText>
-									{cellProps.row.original.market}
-									{cellProps.row.original.isStale && (
-										<ExpiredBadge>
-											{t('futures.market.user.open-orders.badges.expired')}
-										</ExpiredBadge>
-									)}
-								</StyledText>
-								<StyledValue>{cellProps.row.original.orderType}</StyledValue>
-							</MarketContainer>
-						);
-					},
-					sortable: true,
-					width: 50,
-				},
-				{
-					Header: (
-						<StyledTableHeader>{t('futures.market.user.open-orders.table.side')}</StyledTableHeader>
-					),
-					accessor: 'side',
-					Cell: (cellProps: CellProps<any>) => {
-						return (
-							<div>
-								<PositionType side={cellProps.row.original.side} />
-							</div>
-						);
-					},
-					sortable: true,
-					width: 50,
-				},
-				{
-					Header: (
-						<StyledTableHeader>{t('futures.market.user.open-orders.table.size')}</StyledTableHeader>
-					),
-					accessor: 'size',
-					Cell: (cellProps: CellProps<any>) => {
-						return <div>{cellProps.row.original.size}</div>;
-					},
-					sortable: true,
-					width: 50,
-				},
-				{
-					Header: (
-						<StyledTableHeader>
-							{t('futures.market.user.open-orders.table.actions')}
-						</StyledTableHeader>
-					),
-					accessor: 'actions',
-					Cell: (cellProps: CellProps<any>) => {
-						return (
-							<div style={{ display: 'flex' }}>
-								<CancelButton
-									onClick={() => {
-										setAction('cancel');
-									}}
-								>
-									{t('futures.market.user.open-orders.actions.cancel')}
-								</CancelButton>
-								{cellProps.row.original.isExecutable && (
-									<EditButton
-										onClick={() => {
-											setAction('execute');
-										}}
-									>
-										{t('futures.market.user.open-orders.actions.execute')}
-									</EditButton>
-								)}
-							</div>
-						);
-					},
-					width: 50,
-				},
-			]}
-		/>
+		<>
+			<DesktopOnlyView>
+				<StyledTable
+					data={data}
+					highlightRowsOnHover
+					showPagination
+					columns={[
+						{
+							Header: (
+								<StyledTableHeader>
+									{t('futures.market.user.open-orders.table.market-type')}
+								</StyledTableHeader>
+							),
+							accessor: 'market',
+							Cell: (cellProps: CellProps<any>) => {
+								return (
+									<MarketContainer>
+										<IconContainer>
+											<StyledCurrencyIcon currencyKey={cellProps.row.original.marketKey} />
+										</IconContainer>
+										<StyledText>
+											{cellProps.row.original.market}
+											{cellProps.row.original.isStale && (
+												<ExpiredBadge>
+													{t('futures.market.user.open-orders.badges.expired')}
+												</ExpiredBadge>
+											)}
+										</StyledText>
+										<StyledValue>{cellProps.row.original.orderType}</StyledValue>
+									</MarketContainer>
+								);
+							},
+							sortable: true,
+							width: 50,
+						},
+						{
+							Header: (
+								<StyledTableHeader>
+									{t('futures.market.user.open-orders.table.side')}
+								</StyledTableHeader>
+							),
+							accessor: 'side',
+							Cell: (cellProps: CellProps<any>) => {
+								return (
+									<div>
+										<PositionType side={cellProps.row.original.side} />
+									</div>
+								);
+							},
+							sortable: true,
+							width: 50,
+						},
+						{
+							Header: (
+								<StyledTableHeader>
+									{t('futures.market.user.open-orders.table.size')}
+								</StyledTableHeader>
+							),
+							accessor: 'size',
+							Cell: (cellProps: CellProps<any>) => {
+								return <div>{cellProps.row.original.size}</div>;
+							},
+							sortable: true,
+							width: 50,
+						},
+						{
+							Header: (
+								<StyledTableHeader>
+									{t('futures.market.user.open-orders.table.actions')}
+								</StyledTableHeader>
+							),
+							accessor: 'actions',
+							Cell: (cellProps: CellProps<any>) => {
+								return (
+									<div style={{ display: 'flex' }}>
+										<CancelButton
+											onClick={() => {
+												setAction('cancel');
+											}}
+										>
+											{t('futures.market.user.open-orders.actions.cancel')}
+										</CancelButton>
+										{cellProps.row.original.isExecutable && (
+											<EditButton
+												onClick={() => {
+													setAction('execute');
+												}}
+											>
+												{t('futures.market.user.open-orders.actions.execute')}
+											</EditButton>
+										)}
+									</div>
+								);
+							},
+							width: 50,
+						},
+					]}
+				/>
+			</DesktopOnlyView>
+			<MobileOrTabletView>
+				<StyledTable
+					data={data}
+					onTableRowClick={(row) => {
+						setSelectedOrder(row.original);
+					}}
+					columns={[
+						{
+							Header: <StyledTableHeader>Side/Type</StyledTableHeader>,
+							accessor: 'side/type',
+							Cell: (cellProps: CellProps<any>) => (
+								<div>
+									<PositionType side={cellProps.row.original.side} />
+									<div>{cellProps.row.original.orderType}</div>
+								</div>
+							),
+							width: 100,
+						},
+						{
+							Header: <StyledTableHeader>Size</StyledTableHeader>,
+							accessor: 'size',
+							Cell: (cellProps: CellProps<any>) => (
+								<div>
+									{formatCurrency(cellProps.row.original.asset, cellProps.row.original.size, {
+										sign: cellProps.row.original.asset,
+									})}
+								</div>
+							),
+						},
+						{
+							Header: (
+								<StyledTableHeader>
+									{t('futures.market.user.open-orders.table.actions')}
+								</StyledTableHeader>
+							),
+							accessor: 'actions',
+							Cell: (cellProps: CellProps<any>) => {
+								return (
+									<div style={{ display: 'flex' }}>
+										<CancelButton
+											onClick={() => {
+												setAction('cancel');
+											}}
+										>
+											{t('futures.market.user.open-orders.actions.cancel')}
+										</CancelButton>
+										{cellProps.row.original.isExecutable && (
+											<EditButton
+												onClick={() => {
+													setAction('execute');
+												}}
+											>
+												{t('futures.market.user.open-orders.actions.execute')}
+											</EditButton>
+										)}
+									</div>
+								);
+							},
+							width: 100,
+						},
+					]}
+					noResultsMessage={
+						openOrders.length === 0 ? (
+							<TableNoResults>You have no open orders.</TableNoResults>
+						) : undefined
+					}
+				/>
+				<OrderDrawer
+					open={!!selectedOrder}
+					order={selectedOrder}
+					closeDrawer={() => setSelectedOrder(undefined)}
+					setAction={setAction}
+				/>
+			</MobileOrTabletView>
+		</>
 	);
 };
 
@@ -257,6 +343,16 @@ const ExpiredBadge = styled(Badge)`
 	background: ${(props) => props.theme.colors.selectedTheme.red};
 	padding: 1px 5px;
 	line-height: 9px;
+`;
+
+const TableNoResults = styled(GridDivCenteredRow)`
+	padding: 50px 0;
+	justify-content: center;
+	margin-top: -2px;
+	justify-items: center;
+	grid-gap: 10px;
+	color: ${(props) => props.theme.colors.common.primaryWhite};
+	font-size: 16px;
 `;
 
 export default OpenOrdersTable;
